+++ src/components/NavMenu.tsx
@@ -1,10 +1,10 @@
-import { Fragment, useRef, useState } from "react";
+import React, { Fragment, useRef, useState } from "react";
 import { Menu, Transition } from "@headlessui/react";
 import { BsThreeDotsVertical } from "react-icons/bs";
 import Link from "next/link";
 import { useMainContext } from "../MainContext";
 import Toggles from "./settings/Toggles";
-import React from "react";
+import { useWindowWidth } from "@react-hook/window-size/throttled";
 
 function classNames(...classes) {
   return classes.filter(Boolean).join(" ");
@@ -16,7 +16,7 @@ const subOptionStyle =
 const NavMenu = ({ hide = false }) => {
   const context: any = useMainContext();
   const [touched, setTouched] = useState(false);
-
+  const windowWidth = useWindowWidth(); 
   //need to reset wideUI so multi-column (narrow UI doesn't make sense with >1 column) displays properly.
   //Refreshes feed to fix alignment when switching to 1 column and in narrow UI mode
   //Also refreshes feed in mainContext with a useEffect when wideUI is changed if in 1 column mode
@@ -68,47 +68,47 @@ const NavMenu = ({ hide = false }) => {
   const cardStyleRef = useRef<HTMLButtonElement>(null);
 
   return (
-<>
-    <Menu
-      as="div"
-      className="relative flex flex-col items-center flex-grow w-full h-full select-none"
-    >
-      <div className="flex-grow w-full">
-        <Menu.Button
-          aria-label="options"
-          title={"options"}
-          name="Options"
-          className="flex flex-row items-center justify-center w-full h-full border border-transparent rounded-md hover:border-th-border focus:outline-none"
-        >
-          <BsThreeDotsVertical className="flex-none w-5 h-5" />
-        </Menu.Button>
-      </div>
-
-      <Transition
-        as={Fragment}
-        enter="transition ease-out duration-100"
-        enterFrom="transform opacity-0 scale-95"
-        enterTo="transform opacity-100 scale-100"
-        leave="transition ease-in duration-75"
-        leaveFrom="transform opacity-100 scale-100"
-        leaveTo="transform opacity-0 scale-95"
+    <>
+      <Menu
+        as="div"
+        className="relative flex flex-col items-center flex-grow w-full h-full select-none"
       >
-        <Menu.Items
-          className={
-            "absolute right-0 w-40 mt-10 origin-top-right bg-th-background2 rounded-md shadow-lg focus:outline-none border border-th-border ring-1 ring-th-base ring-opacity-5 z-50  " +
-            (hide && " hidden")
-          }
+        <div className="flex-grow w-full">
+          <Menu.Button
+            aria-label="options"
+            title={"options"}
+            name="Options"
+            className="flex flex-row items-center justify-center w-full h-full border border-transparent rounded-md hover:border-th-border focus:outline-none"
+          >
+            <BsThreeDotsVertical className="flex-none w-5 h-5" />
+          </Menu.Button>
+        </div>
+
+        <Transition
+          as={Fragment}
+          enter="transition ease-out duration-100"
+          enterFrom="transform opacity-0 scale-95"
+          enterTo="transform opacity-100 scale-100"
+          leave="transition ease-in duration-75"
+          leaveFrom="transform opacity-100 scale-100"
+          leaveTo="transform opacity-0 scale-95"
         >
-          <div className="py-1">
-            <Menu.Item>
-              {({ active }) => (
-                <div className="group">
-                  <div
-                    className={classNames(
-                      active ? "bg-th-highlight " : "",
-                      "block text-sm"
-                    )}
-                  >
+          <Menu.Items
+            className={
+              "absolute right-0 w-40 mt-10 origin-top-right bg-th-background2 rounded-md shadow-lg focus:outline-none border border-th-border ring-1 ring-th-base ring-opacity-5 z-50  " +
+              (hide && " hidden")
+            }
+          >
+            <div className="py-1">
+              <Menu.Item>
+                {({ active }) => (
+                  <div className="group">
+                    <div
+                      className={classNames(
+                        active ? "bg-th-highlight " : "",
+                        "block text-sm"
+                      )}
+                    >
                       <button
                         className="flex flex-row items-center justify-center w-full h-6 px-4 py-5"
                         onTouchStart={() => setTouched(true)}
@@ -124,120 +124,121 @@ const NavMenu = ({ hide = false }) => {
                       >
                         Column Count
                       </button>
+                    </div>
+                    <ul
+                      className={
+                        (active && !touched ? "block " : "hidden ") +
+                        (!touched
+                          ? " group-hover:block group-focus:block "
+                          : "") +
+                        "absolute top-0 w-32 -left-32  bg-th-background2 rounded-md shadow-lg border border-th-border text-right"
+                      }
+                    >
+                      <li>
+                        <div
+                          className={
+                            (context.columnOverride === 0 ? "font-bold " : "") +
+                            " px-4 py-3 text-sm hover:bg-th-highlight   mt-1 cursor-pointer"
+                          }
+                          onClick={(e) => {
+                            setColumnCount(0);
+                          }}
+                        >
+                          Automatic
+                        </div>
+                      </li>
+                      <li>
+                        <div
+                          className={
+                            (context.columnOverride === 1 ? "font-bold " : "") +
+                            ` ${subOptionStyle}`
+                          }
+                          onClick={(e) => {
+                            setColumnCount(1);
+                          }}
+                        >
+                          One
+                        </div>
+                      </li>
+                      <li>
+                        <div
+                          className={
+                            (context.columnOverride === 2 ? "font-bold " : "") +
+                            ` ${subOptionStyle}`
+                          }
+                          onClick={(e) => {
+                            setColumnCount(2);
+                          }}
+                        >
+                          Two
+                        </div>
+                      </li>
+                      <li>
+                        <div
+                          className={
+                            (context.columnOverride === 3 ? "font-bold " : "") +
+                            ` ${subOptionStyle}`
+                          }
+                          onClick={(e) => {
+                            setColumnCount(3);
+                          }}
+                        >
+                          Three
+                        </div>
+                      </li>
+                      <li>
+                        <div
+                          className={
+                            (context.columnOverride === 4 ? "font-bold " : "") +
+                            ` ${subOptionStyle}`
+                          }
+                          onClick={(e) => {
+                            setColumnCount(4);
+                          }}
+                        >
+                          Four
+                        </div>
+                      </li>
+                      <li>
+                        <div
+                          className={
+                            (context.columnOverride === 5 ? "font-bold " : "") +
+                            ` ${subOptionStyle}`
+                          }
+                          onClick={(e) => {
+                            setColumnCount(5);
+                          }}
+                        >
+                          Five
+                        </div>
+                      </li>
+                      <li>
+                        <div
+                          className={
+                            (context.columnOverride === 7 ? "font-bold " : "") +
+                            ` ${subOptionStyle}`
+                          }
+                          onClick={(e) => {
+                            setColumnCount(7);
+                          }}
+                        >
+                          Seven
+                        </div>
+                      </li>
+                    </ul>
                   </div>
-                  <ul
-                    className={
-                      (active && !touched ? "block " : "hidden ") +
-                      (!touched ? " group-hover:block group-focus:block " : "") + 
-                      "absolute top-0 w-32 -left-32  bg-th-background2 rounded-md shadow-lg border border-th-border text-right"
-                    }
-                  >
-                    <li>
-                      <div
-                        className={
-                          (context.columnOverride === 0 ? "font-bold " : "") +
-                          " px-4 py-3 text-sm hover:bg-th-highlight   mt-1 cursor-pointer"
-                        }
-                        onClick={(e) => {
-                          setColumnCount(0);
-                        }}
-                      >
-                        Automatic
-                      </div>
-                    </li>
-                    <li>
-                      <div
-                        className={
-                          (context.columnOverride === 1 ? "font-bold " : "") +
-                          ` ${subOptionStyle}`
-                        }
-                        onClick={(e) => {
-                          setColumnCount(1);
-                        }}
-                      >
-                        One
-                      </div>
-                    </li>
-                    <li>
-                      <div
-                        className={
-                          (context.columnOverride === 2 ? "font-bold " : "") +
-                          ` ${subOptionStyle}`
-                        }
-                        onClick={(e) => {
-                          setColumnCount(2);
-                        }}
-                      >
-                        Two
-                      </div>
-                    </li>
-                    <li>
-                      <div
-                        className={
-                          (context.columnOverride === 3 ? "font-bold " : "") +
-                          ` ${subOptionStyle}`
-                        }
-                        onClick={(e) => {
-                          setColumnCount(3);
-                        }}
-                      >
-                        Three
-                      </div>
-                    </li>
-                    <li>
-                      <div
-                        className={
-                          (context.columnOverride === 4 ? "font-bold " : "") +
-                          ` ${subOptionStyle}`
-                        }
-                        onClick={(e) => {
-                          setColumnCount(4);
-                        }}
-                      >
-                        Four
-                      </div>
-                    </li>
-                    <li>
-                      <div
-                        className={
-                          (context.columnOverride === 5 ? "font-bold " : "") +
-                          ` ${subOptionStyle}`
-                        }
-                        onClick={(e) => {
-                          setColumnCount(5);
-                        }}
-                      >
-                        Five
-                      </div>
-                    </li>
-                    <li>
-                      <div
-                        className={
-                          (context.columnOverride === 7 ? "font-bold " : "") +
-                          ` ${subOptionStyle}`
-                        }
-                        onClick={(e) => {
-                          setColumnCount(7);
-                        }}
-                      >
-                        Seven
-                      </div>
-                    </li>
-                  </ul>
-                </div>
-              )}
-            </Menu.Item>
-            <Menu.Item>
-              {({ active }) => (
-                <div className="group">
-                  <div
-                    className={classNames(
-                      active ? "bg-th-highlight " : "",
-                      "block text-sm"
-                    )}
-                  
-                  >
-                     <button
+                )}
+              </Menu.Item>
+              <Menu.Item>
+                {({ active }) => (
+                  <div className="group">
+                    <div
+                      className={classNames(
+                        active ? "bg-th-highlight " : "",
+                        "block text-sm"
+                      )}
+                    >
+                      <button
                         className="flex flex-row items-center justify-center w-full h-6 px-4 py-5 "
                         onTouchStart={() => setTouched(true)}
                         onClick={(e) => {
@@ -252,85 +253,108 @@ const NavMenu = ({ hide = false }) => {
                       >
                         Card Style
                       </button>
+                    </div>
+                    <ul
+                      className={
+                        (active && !touched ? "block " : "hidden ") +
+                        (!touched
+                          ? " group-hover:block group-focus:block "
+                          : "") +
+                        " absolute top-0 w-32 -left-32  bg-th-background2 rounded-md shadow-lg border border-th-border text-right"
+                      }
+                    >
+                      <li>
+                        <div
+                          className={
+                            (context.cardStyle === "card1" && !context.mediaOnly
+                              ? " font-bold "
+                              : "") + ` ${subOptionStyle} mt-1`
+                          }
+                          onClick={(e) => {
+                            setCardStyle("card1");
+                            // if(context.columnOverride == 1 || context.columns == 1) {context.setColumnOverride(0);};
+
+                            context.setMediaOnly(false);
+                          }}
+                        >
+                          Original Card
+                        </div>
+                      </li>
+                      <li>
+                        <div
+                          className={
+                            (context.cardStyle === "card2"
+                              ? " font-bold "
+                              : "") + ` ${subOptionStyle}`
+                          }
+                          onClick={(e) => {
+                            setCardStyle("card2");
+                            //if(context.columnOverride == 1 || context.columns == 1) {context.setColumnOverride(0);};
+
+                            context.setMediaOnly(false);
+                          }}
+                        >
+                          Compact Card
+                        </div>
+                      </li>
+                      <li>
+                        <div
+                          className={
+                            (context.cardStyle == "card1" && context.mediaOnly
+                              ? " font-bold "
+                              : "") + ` ${subOptionStyle}`
+                          }
+                          onClick={(e) => {
+                            setCardStyle("card1");
+                            //if(context.columnOverride == 1 || context.columns == 1) {context.setColumnOverride(0);};
+                            context.setMediaOnly(true);
+                          }}
+                        >
+                          Media Card
+                        </div>
+                      </li>
+                      <li>
+                        <div
+                          className={
+                            (context.cardStyle == "row1" ? " font-bold " : "") +
+                            ` ${subOptionStyle}`
+                          }
+                          onClick={(e) => {
+                            setCardStyle("row1");
+                            context.setMediaOnly(false);
+                            context.setColumnOverride(1);
+                          }}
+                        >
+                          Classic Rows
+                        </div>
+                      </li>
+                    </ul>
                   </div>
-                  <ul
-                    className={
-                      (active && !touched ? "block " : "hidden ") +
-                      (!touched ? " group-hover:block group-focus:block " : "") + 
-                      " absolute top-0 w-32 -left-32  bg-th-background2 rounded-md shadow-lg border border-th-border text-right"
-                    }
-                  >
-                    <li>
-                      <div
-                        className={
-                          (context.cardStyle === "card1" && !context.mediaOnly
-                            ? " font-bold "
-                            : "") + ` ${subOptionStyle} mt-1`
-                        }
-                        onClick={(e) => {
-                          setCardStyle("card1");
-                          // if(context.columnOverride == 1 || context.columns == 1) {context.setColumnOverride(0);};
+                )}
+              </Menu.Item>
 
-                          context.setMediaOnly(false);
-                        }}
-                      >
-                        Original Card
-                      </div>
-                    </li>
-                    <li>
-                      <div
-                        className={
-                          (context.cardStyle === "card2" ? " font-bold " : "") +
-                          ` ${subOptionStyle}`
-                        }
-                        onClick={(e) => {
-                          setCardStyle("card2");
-                          //if(context.columnOverride == 1 || context.columns == 1) {context.setColumnOverride(0);};
-
-                          context.setMediaOnly(false);
-                        }}
-                      >
-                        Compact Card
-                      </div>
-                    </li>
-                    <li>
-                      <div
-                        className={
-                          (context.cardStyle == "card1" && context.mediaOnly
-                            ? " font-bold "
-                            : "") + ` ${subOptionStyle}`
-                        }
-                        onClick={(e) => {
-                          setCardStyle("card1");
-                          //if(context.columnOverride == 1 || context.columns == 1) {context.setColumnOverride(0);};
-                          context.setMediaOnly(true);
-                        }}
-                      >
-                        Media Card
-                      </div>
-                    </li>
-                    <li>
-                      <div
-                        className={
-                          (context.cardStyle == "row1" ? " font-bold " : "") +
-                          ` ${subOptionStyle}`
-                        }
-                        onClick={(e) => {
-                          setCardStyle("row1");
-                          context.setMediaOnly(false);
-                          context.setColumnOverride(1);
-                        }}
-                      >
-                        Classic Rows
-                      </div>
-                    </li>
-                  </ul>
-                </div>
+              {windowWidth > 640 && (
+                <>
+                  {["hoverplay"].map((setting: any) => (
+                    <Menu.Item key={setting}>
+                      {({ active }) => (
+                        <div
+                          className={classNames(
+                            active ? "bg-th-highlight " : "",
+                            "block px-4 text-sm "
+                          )}
+                        >
+                          <Toggles
+                            setting={setting}
+                            externalStyles="py-2 cursor-pointer"
+                          />
+                        </div>
+                      )}
+                    </Menu.Item>
+                  ))}
+                </>
               )}
-            </Menu.Item>
-
-            {["hoverplay", "autoplay", "audioOnHover", "theme"].map(
-              (setting: any) => (
+              {["autoplay", "audioOnHover", "theme"].map((setting: any) => (
                 <Menu.Item key={setting}>
                   {({ active }) => (
                     <div
@@ -346,9 +370,8 @@ const NavMenu = ({ hide = false }) => {
                     </div>
                   )}
                 </Menu.Item>
-              )
-            )}
-               {context.cardStyle === "card1" && context.mediaOnly && (
+              ))}
+              {context.cardStyle === "card1" && context.mediaOnly && (
                 <>
                   {["uniformHeights"].map((setting: any) => (
                     <Menu.Item key={setting}>
@@ -369,43 +392,43 @@ const NavMenu = ({ hide = false }) => {
                   ))}
                 </>
               )}
-            <Menu.Item>
-              {({ active }) => (
-                <div
-                  className={classNames(
-                    active ? "bg-th-highlight " : "",
-                    "block px-4  text-sm"
-                  )}
-                >
-                  <Link href="/settings" passHref={true}>
-                    <a className="flex flex-row justify-center py-2 cursor-pointer select-none">
-                      Settings
-                    </a>
-                  </Link>
-                </div>
-              )}
-            </Menu.Item>
-            <Menu.Item>
-              {({ active }) => (
-                <div
-                  className={classNames(
-                    active ? "bg-th-highlight " : "",
-                    "block px-4  text-sm"
-                  )}
-                >
-                  <Link href="/about" passHref={true}>
-                    <a className="flex flex-row justify-center py-2 cursor-pointer select-none">
-                      About
-                    </a>
-                  </Link>
-                </div>
-              )}
-            </Menu.Item>
-          </div>
-        </Menu.Items>
-      </Transition>
-    </Menu>
-    <Menu as={"div"} className={"relative font-normal"}>
+              <Menu.Item>
+                {({ active }) => (
+                  <div
+                    className={classNames(
+                      active ? "bg-th-highlight " : "",
+                      "block px-4  text-sm"
+                    )}
+                  >
+                    <Link href="/settings" passHref={true}>
+                      <a className="flex flex-row justify-center py-2 cursor-pointer select-none">
+                        Settings
+                      </a>
+                    </Link>
+                  </div>
+                )}
+              </Menu.Item>
+              <Menu.Item>
+                {({ active }) => (
+                  <div
+                    className={classNames(
+                      active ? "bg-th-highlight " : "",
+                      "block px-4  text-sm"
+                    )}
+                  >
+                    <Link href="/about" passHref={true}>
+                      <a className="flex flex-row justify-center py-2 cursor-pointer select-none">
+                        About
+                      </a>
+                    </Link>
+                  </div>
+                )}
+              </Menu.Item>
+            </div>
+          </Menu.Items>
+        </Transition>
+      </Menu>
+      <Menu as={"div"} className={"relative font-normal"}>
         <Menu.Button
           aria-label="column count options"
           ref={columnMenuRef}
@@ -430,7 +453,9 @@ const NavMenu = ({ hide = false }) => {
             <Menu.Item as={"li"}>
               <div
                 className={
-                  (context.columnOverride === 0 ? "font-bold bg-th-highlight " : "") +
+                  (context.columnOverride === 0
+                    ? "font-bold bg-th-highlight "
+                    : "") +
                   " px-4 py-3 text-sm hover:bg-th-highlight   mt-1 cursor-pointer"
                 }
                 onClick={(e) => {
@@ -443,8 +468,9 @@ const NavMenu = ({ hide = false }) => {
             <Menu.Item as={"li"}>
               <div
                 className={
-                  (context.columnOverride === 1 ? "font-bold bg-th-highlight " : "") +
-                  ` ${subOptionStyle}`
+                  (context.columnOverride === 1
+                    ? "font-bold bg-th-highlight "
+                    : "") + ` ${subOptionStyle}`
                 }
                 onClick={(e) => {
                   setColumnCount(1);
@@ -456,8 +482,9 @@ const NavMenu = ({ hide = false }) => {
             <Menu.Item as={"li"}>
               <div
                 className={
-                  (context.columnOverride === 2 ? "font-bold bg-th-highlight " : "") +
-                  ` ${subOptionStyle}`
+                  (context.columnOverride === 2
+                    ? "font-bold bg-th-highlight "
+                    : "") + ` ${subOptionStyle}`
                 }
                 onClick={(e) => {
                   setColumnCount(2);
@@ -469,8 +496,9 @@ const NavMenu = ({ hide = false }) => {
             <Menu.Item as={"li"}>
               <div
                 className={
-                  (context.columnOverride === 3 ? "font-bold bg-th-highlight " : "") +
-                  ` ${subOptionStyle}`
+                  (context.columnOverride === 3
+                    ? "font-bold bg-th-highlight "
+                    : "") + ` ${subOptionStyle}`
                 }
                 onClick={(e) => {
                   setColumnCount(3);
@@ -482,8 +510,9 @@ const NavMenu = ({ hide = false }) => {
             <Menu.Item as={"li"}>
               <div
                 className={
-                  (context.columnOverride === 4 ? "font-bold bg-th-highlight " : "") +
-                  ` ${subOptionStyle}`
+                  (context.columnOverride === 4
+                    ? "font-bold bg-th-highlight "
+                    : "") + ` ${subOptionStyle}`
                 }
                 onClick={(e) => {
                   setColumnCount(4);
@@ -495,8 +524,9 @@ const NavMenu = ({ hide = false }) => {
             <Menu.Item as={"li"}>
               <div
                 className={
-                  (context.columnOverride === 5 ? "font-bold bg-th-highlight " : "") +
-                  ` ${subOptionStyle}`
+                  (context.columnOverride === 5
+                    ? "font-bold bg-th-highlight "
+                    : "") + ` ${subOptionStyle}`
                 }
                 onClick={(e) => {
                   setColumnCount(5);
@@ -508,8 +538,9 @@ const NavMenu = ({ hide = false }) => {
             <Menu.Item as={"li"}>
               <div
                 className={
-                  (context.columnOverride === 7 ? "font-bold bg-th-highlight " : "") +
-                  ` ${subOptionStyle}`
+                  (context.columnOverride === 7
+                    ? "font-bold bg-th-highlight "
+                    : "") + ` ${subOptionStyle}`
                 }
                 onClick={(e) => {
                   setColumnCount(7);
@@ -546,7 +577,9 @@ const NavMenu = ({ hide = false }) => {
             <Menu.Item as={"li"}>
               <div
                 className={
-                  ((context.cardStyle === "card1" || context.cardStyle === "default") && !context.mediaOnly
+                  ((context.cardStyle === "card1" ||
+                    context.cardStyle === "default") &&
+                  !context.mediaOnly
                     ? " font-bold bg-th-highlight "
                     : "") + ` ${subOptionStyle} mt-1`
                 }
@@ -563,8 +596,9 @@ const NavMenu = ({ hide = false }) => {
             <Menu.Item as={"li"}>
               <div
                 className={
-                  (context.cardStyle === "card2" ? " font-bold bg-th-highlight " : "") +
-                  ` ${subOptionStyle}`
+                  (context.cardStyle === "card2"
+                    ? " font-bold bg-th-highlight "
+                    : "") + ` ${subOptionStyle}`
                 }
                 onClick={(e) => {
                   setCardStyle("card2");
@@ -595,8 +629,9 @@ const NavMenu = ({ hide = false }) => {
             <Menu.Item as={"li"}>
               <div
                 className={
-                  (context.cardStyle == "row1" ? " font-bold bg-th-highlight " : "") +
-                  ` ${subOptionStyle}`
+                  (context.cardStyle == "row1"
+                    ? " font-bold bg-th-highlight "
+                    : "") + ` ${subOptionStyle}`
                 }
                 onClick={(e) => {
                   setCardStyle("row1");
